improve stacktraces on windows and linux
